+++ backend/typst_writer/src/colors.rs
@@ -23,17 +23,20 @@ impl Display for Color {
 
 pub fn get_color_preamble(has_color: bool) -> String {
     let colored: Color;
+    let solution_background: Color;
     // Graphing colors
     let primary: Color;
     let secondary: Color;
     let tertiary: Color;
     if has_color {
         colored = Color::new(22, 10, 33); // Purple
+        solution_background = Color::new(98, 96, 92); // Light yellow
         primary = Color::new(9, 3, 18); // Dark purple
         secondary = colored.clone();
         tertiary = Color::new(30, 23, 39); // Light purple
     } else {
         colored = Color::new(10, 10, 10); // Gray
+        solution_background = Color::new(88, 88, 88); // Light gray
         primary = Color::new(0, 0, 0); // Black
         secondary = Color::new(8, 8, 8); // Gray?
         tertiary = Color::new(16, 16, 16); // Grayer?
@@ -42,6 +45,7 @@ pub fn get_color_preamble(has_color: bool) -> String {
     format!(
         "
 #let primary_color = color.linear-rgb({primary})
+#let solution_color = color.linear-rgb({solution_background})
 #let secondary_color = color.linear-rgb({secondary})
 #let tertiary_color = color.linear-rgb({tertiary})
 #let colored(x) = text(fill: color.linear-rgb({colored}), $#x$)