+++ backend/typst_writer/src/preamble.rs
@@ -1,4 +1,4 @@
-pub static PREAMBLE_STR: &str = r#"
+pub static PREAMBLE_STR: &str = r##"
 #import "/preamble.typ": *
 #import "@preview/equate:0.3.2": equate, share-align
 #import "@preview/zero:0.5.0": num, set-num
@@ -13,4 +13,30 @@ pub static PREAMBLE_STR: &str = r#"
 })
 #set-num(decimal-separator: ",")
 #let item(content) = block(breakable: false, content)
-"#;
+#let equation-solution(equations, operations) = {
+  context {
+    let max-eq-width = 0pt
+    let max-op-width = 0pt
+
+    for eq in equations {
+      let size = measure(eq)
+      if size.width > max-eq-width {
+        max-eq-width = size.width
+      }
+    }
+
+    let color-operations = operations.map(op => if op != $$ { colored(op) } else { op })
+
+    share-align({
+      grid(
+        columns: (max-eq-width, auto),
+        inset: (top: 0.2em, rest: 0.5em),
+        align: (left, horizon + left),
+        grid.vline(x: 1, stroke: (paint: line_color, thickness: 0.5pt)),
+        ..equations.zip(color-operations).flatten(),
+      )
+    })
+  }
+}
+
+"##;